pyproject.toml
[tool.poetry]
name = "PCParadisebot"
version = "1.0.0"
description = "A small discord bot for pcparadise"
# all the people who've developed on this project can be found
# on the github repo, only the original author is written below.
authors = ["Arvinder Dhanoa <awsomearvinder@gmail.com>"]
maintainers = [ "Ganoosh <swajabbadaja@gmail.com>" ]
license = "GPLv3"
packages = [
  { include = "src" }
]

[tool.poetry.dependencies]
"discord.py" = "^2.0"
appdirs = "^1.4.4"
python = "^3.10"
aiosqlite = "^0.17.0"
pytimeparse = "^1.1.8"
tortoise-orm = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "22.3.0"
pylint = "^2.14.5"
mypy = "~0.942"

[tool.poetry.scripts]
run = "src.main:main"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.core.masonry.api"
